# server/openapi.yaml
openapi: 3.0.0
info:
  title: Mini Supermarket API
  version: 1.0.0
  description: API cho hệ thống quản lý siêu thị mini
servers:
  - url: 'http://localhost:5000'
...
paths:
...
  /api/staff/:
    get:
      tags:
        - staff
      summary: get All Staff
      operationId: getAllStaff
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
    post:
      tags:
        - staff
      summary: create Staff
      operationId: createStaff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Tạo thành công
        '400':
          description: Dữ liệu không hợp lệ
  /api/staff/stats:
    get:
      tags:
        - staff
      summary: get Staff Stats
      operationId: getStaffStats
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
  '/api/staff/{id}':
    get:
      tags:
        - staff
      summary: get Staff By Id
      operationId: getStaffById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
    put:
      tags:
        - staff
      summary: update Staff
      operationId: updateStaff
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
    delete:
      tags:
        - staff
      summary: delete Staff
      operationId: deleteStaff
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
  '/api/staff/{id}/permanent':
    delete:
      tags:
        - staff
      summary: permanent Delete Staff
      operationId: permanentDeleteStaff
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
  '/api/staff/{id}/activate':
    patch:
      tags:
        - staff
      summary: activate Staff
      operationId: activateStaff
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
  '/api/staff/account/{accountId}':
    get:
      tags:
        - staff
      summary: get Staff By Account Id
      operationId: getStaffByAccountId
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
...
  /api/telemetry/traces:
    get:
      tags:
        - telemetry
      summary: get Traces
      operationId: getTraces
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
    delete:
      tags:
        - telemetry
      summary: clear Traces
      operationId: clearTraces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Thành công
        '400':
          description: Dữ liệu không hợp lệ
...